+++ apps/client/src/app/pages/blog/2022/07/ghostfolio-meets-internet-identity/ghostfolio-meets-internet-identity-page.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  host: { class: 'page' },
+  selector: 'gf-ghostfolio-meets-internet-identity-page',
+  styleUrls: ['./ghostfolio-meets-internet-identity-page.scss'],
+  templateUrl: './ghostfolio-meets-internet-identity-page.html'
+})
+export class GhostfolioMeetsInternetIdentityPageComponent {}